drivers: rtc: RV8263 get returns error when OS flag is set
A get time API call returns an error when the OS flag is
set to indicate an invalid state. Also clear the
flag once during the initialization.

+++ drivers/rtc/rtc_rv8263.c
@@ -281,7 +281,7 @@ static int rv8263c8_time_get(const struct device *dev, struct rtc_time *timeptr)
 static int rv8263c8_init(const struct device *dev)
 {
 	int err;
-	int temp;
+	uint8_t temp;
 	struct rv8263c8_data *data = dev->data;
 	const struct rv8263c8_config *config = dev->config;
